refactor(yabloc_common): apply static analysis (#7481)
* refactor based on linter


* restore unwanted change


* remove unnecessary comment


* style(pre-commit): autofix

* change to const double


* add static cast


---------

+++ localization/yabloc/yabloc_common/include/yabloc_common/ground_server/filter/moving_averaging.hpp
@@ -32,7 +32,7 @@ public:
 
     Eigen::Vector3f mean = Eigen::Vector3f::Zero();
     for (const Eigen::Vector3f & v : buffer_) mean += v;
-    mean /= buffer_.size();
+    mean /= static_cast<float>(buffer_.size());
 
     return mean.normalized();
   }
